requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.1
django-phonenumber-field==7.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
phonenumberslite==8.12.56
psycopg2-binary==2.9.3
PyJWT==2.5.0
pytz==2022.2.1
sqlparse==0.4.3
